HCK-9614: adapt FE to work in the browser
